redesign components structure
