+++ configs/webpack.common.js
@@ -4,6 +4,10 @@ const HTMLWebpackPlugin = require('html-webpack-plugin');
 const source = path.resolve(__dirname, '../src');
 const outputPath = path.resolve(__dirname, '../dist');
 const entryPath = path.resolve(__dirname, '../src/index.jsx');
+const crypto = require('crypto');
+const crypto_orig_createHash = crypto.createHash;
+crypto.createHash = (algorithm) =>
+  crypto_orig_createHash(algorithm == 'md4' ? 'sha256' : algorithm);
 
 module.exports = {
   entry: [entryPath],
