+++ src/App/App.css
@@ -1,4 +1,15 @@
 @import url('https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap');
+body {
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu',
+    'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New', monospace;
+}
 
 html {
   box-sizing: border-box;
@@ -24,12 +35,8 @@ body {
 main {
   background: url(../assets/images/background_main.svg) no-repeat;
   background-size: cover;
-}
-
-.main__wrapper {
-  margin: 0px auto;
-  max-width: 1440px;
   height: 100vh;
+  margin: 0px auto;
 }
 
 .main__title {
@@ -282,16 +289,20 @@ main {
   font-size: 12px;
   border: 1px solid #3077c6;
 }
+.market__links {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  margin: 0px auto;
+  padding-bottom: 120px;
+}
 
 .recommend {
   background-color: #eaf0f9;
+  height: 100vh;
+  padding: 0px 104px;
 }
 
-.recommend__wrapper {
-  padding: 0px 104px;
-  margin: 0px auto;
-  max-width: 1440px;
-}
 .recommend__title {
   font-family: 'Roboto';
   font-style: normal;
@@ -305,18 +316,11 @@ main {
   color: #383838;
 }
 
-.market__links {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  margin: 0px auto;
-  padding-bottom: 120px;
-}
-
-/* слайдер */
 .recommend-list__wrapper {
   overflow: hidden;
   padding-bottom: 120px;
+  max-width: 1440px;
+  margin: 0px auto;
 }
 .recommend__list {
   display: flex;
@@ -332,8 +336,8 @@ main {
 }
 
 .recommend__list li img {
-  width: 276px;
-  height: 276px;
+  width: 20rem;
+  height: 20rem;
 }
 
 .recommend__list li a {
@@ -356,6 +360,7 @@ main {
   color: #bfbfbf;
 }
 
+/*
 .slider-buttons {
   text-align: center;
   display: flex;
@@ -364,7 +369,8 @@ main {
   position: relative;
   top: 150px;
 }
-.btn-prev {
+
+ .btn-prev {
   position: relative;
   z-index: 1;
   background-color: transparent;
@@ -384,16 +390,4 @@ main {
   border: none;
   background-image: url(../assets/images/svg/arrow.svg);
   background-size: 40px;
-}
-
-.not_found {
-  position: relative;
-  top: 300px;
-  bottom: 300px;
-  right: 0px;
-  left: 0px;
-  display: flex;
-  justify-content: center;
-  color: black;
-  font-size: 60px;
-}
+} */
