+++ src/components/Header/Index.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import Logo from './Logo';
+import Navigation from './Navigation';
+
+export default function Header() {
+  return (
+    <header className='header header__wrapper'>
+      <Logo />
+      <Navigation />
+    </header>
+  );
+}
