+++ src/components/Header/Logo.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function Header() {
+  return (
+    <div className='header__logo'>
+      <a href='./' aria-label='./' className='header__logo-link' />
+    </div>
+  );
+}
